requirements.txt
pika
python-pushover
flask
pyserial
bluepy